Refactor s3fs into a basefs class which is independent from the actual S3 interface.
